pyproject.toml
[build-system]
requires = [ "setuptools",]
build-backend = "setuptools.build_meta"

[project]
requires-python = "==3.9.22"
name = "topical_authority_discovery"
readme = "README.md"
dynamic = [ "version",]
dependencies = [
 # Core Kedro and interactive environment
 "ipython>=8.10",
 "jupyterlab>=3.0", # Kept as per instruction
 "notebook", # Kept as per instruction
 "kedro[jupyter]~=0.19.12",
 "kedro-viz>=6.7.0",
 "kedro-datasets[ibis-duckdb,ibis-polars,ibis-sqlite,networkx,pandas-csvdataset,yaml]~=4.1.0", # Updated and kept as per instruction
 # Data processing and Machine Learning
 "pandas~=2.2.3",
 "numpy~=2.0.2",
 "scikit-learn~=1.6.1", # Kept as per instruction
 # Data formats
 "pyarrow~=20.0.0", # Updated (was >=3.1.5) and kept as per instruction
 # NLP
 "spacy~=3.8.5",
 "en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85",
 "pytextrank~=3.3.0",
 "ibis-framework[bigquery,duckdb,examples]>=10.5.0",
 "google-cloud-bigquery[bqstorage]>=3.34.0",
]

[project.scripts]
topical-authority-discovery = "topical_authority_discovery.__main__:main"

[tool.kedro]
package_name = "topical_authority_discovery"
project_name = "topical-authority-discovery"
kedro_init_version = "0.19.12"
tools = "['Custom Logging', 'Data Structure']"
example_pipeline = "True"
source_dir = "src"

[project.entry-points."kedro.hooks"]

[tool.setuptools.dynamic.version]
attr = "topical_authority_discovery.__version__"

[tool.setuptools.packages.find]
where = [ "src",]
namespaces = false

[dependency-groups]
dev = [
    "gensim>=4.3.2",
    "kaggle>=1.7.4.5",
    "pip>=25.1.1",
    "pip-tools>=7.4.1",
    "pyldavis>=3.4.1",
    "pytest>=8.3.5",
]
